replace manual code with generated code for pAddURShiftThenLShiftToRRotation_pAddURShiftThenLShiftToRRotationSym.

+++ src/hotspot/share/opto/addnode.cpp
@@ -415,7 +415,7 @@ if (_JOG_in1->Opcode() == Op_URShiftI
     && _JOG_in11 == _JOG_in21) {
 jint lshift = phase->type(_JOG_in22)->isa_int()->get_con();
 jint rshift = phase->type(_JOG_in12)->isa_int()->get_con();
-if ((lshift & 31) == (java_subtract(32, (rshift & 31)))) {
+if ((lshift & 31) == (java_subtract((jint) 32, (rshift & 31)))) {
 return new RotateRightNode(_JOG_in11, phase->intcon(rshift & 31), TypeInt::INT);
 }
 }
@@ -436,7 +436,7 @@ if (_JOG_in1->Opcode() == Op_LShiftI
     && _JOG_in11 == _JOG_in21) {
 jint lshift = phase->type(_JOG_in12)->isa_int()->get_con();
 jint rshift = phase->type(_JOG_in22)->isa_int()->get_con();
-if ((lshift & 31) == (java_subtract(32, (rshift & 31)))) {
+if ((lshift & 31) == (java_subtract((jint) 32, (rshift & 31)))) {
 return new RotateRightNode(_JOG_in11, phase->intcon(rshift & 31), TypeInt::INT);
 }
 }
